squash! Add tests for LTTBAggregation
Use executeAggregation for tests instead of manual iterate/newState
tests

executeAggregation from the test framework covers many scenarios that
can be applied here

Also, adapt tests to incorporate changes from code review

Rename to LTTBAggregation

Adapt tests to access timestamp and value on the `Point` record.
Previously, `Point` was a class
