+++ extensions/functions/src/test/java/io/crate/operation/aggregation/LttbStateTest.java
@@ -23,10 +23,8 @@ package io.crate.operation.aggregation;
 
 import static io.crate.testing.TestingHelpers.createNodeContext;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.ArrayList;
-import java.util.List;
 
 import org.apache.lucene.util.RamUsageEstimator;
 import org.elasticsearch.common.io.stream.BytesStreamOutput;
@@ -35,8 +33,7 @@ import org.junit.Before;
 import org.junit.Test;
 
 import io.crate.data.breaker.RamAccounting;
-import io.crate.operation.aggregation.LargestTriangleThreeBucketsAggregation.LttbState;
-import io.crate.operation.aggregation.LargestTriangleThreeBucketsAggregation.Point;
+import io.crate.operation.aggregation.LTTBAggregation.LttbState;
 import io.crate.testing.PlainRamAccounting;
 
 public class LttbStateTest extends AggregationTestCase {
@@ -44,118 +41,6 @@ public class LttbStateTest extends AggregationTestCase {
     @Before
     public void prepareFunctions() {
         nodeCtx = createNodeContext(null, null);
-    }
-
-    @Test
-    public void test_should_sort_x_value_when_downsampling() throws Exception {
-        LttbState state = new LttbState();
-        RamAccounting ramAccounting = new PlainRamAccounting();
-
-        state.init(ramAccounting, 3);
-
-        state.addPoint(ramAccounting, 3L, 30.0);
-        state.addPoint(ramAccounting, 2L, 20.0);
-        state.addPoint(ramAccounting, 1L, 10.0);
-
-        List<Point> result = state.downsample();
-        assertThat(result).extracting(i -> i.timestamp).containsExactly(1L, 2L, 3L);
-    }
-
-    @Test
-    public void test_should_downsample_to_threshold_number_of_points() throws Exception {
-        LttbState state = new LttbState();
-        RamAccounting ramAccounting = new PlainRamAccounting();
-        state.init(ramAccounting, 3);
-
-        state.addPoint(ramAccounting, 10L, 100.0);
-        state.addPoint(ramAccounting, 9L, 20.0);
-        state.addPoint(ramAccounting, 8L, 5.0);
-        state.addPoint(ramAccounting, 7L, 10.0);
-        state.addPoint(ramAccounting, 6L, 50.0);
-        state.addPoint(ramAccounting, 5L, 20.0);
-        state.addPoint(ramAccounting, 4L, 10.0);
-        state.addPoint(ramAccounting, 3L, 20.0);
-        state.addPoint(ramAccounting, 2L, 80.0);
-        state.addPoint(ramAccounting, 1L, 90.0);
-
-        assertThat(state.getPoints()).hasSize(10);
-
-        List<Point> result = state.downsample();
-        assertThat(result).extracting(i -> i.timestamp).hasSize(3);
-    }
-
-    @Test
-    public void test_should_select_the_points_that_form_largest_area_when_downsampling() throws Exception {
-        LttbState state = new LttbState();
-        RamAccounting ramAccounting = new PlainRamAccounting();
-        state.init(ramAccounting, 3);
-
-        state.addPoint(ramAccounting, 1L, 10.0);
-        state.addPoint(ramAccounting, 2L, 12.0);
-        state.addPoint(ramAccounting, 3L, 50.0);
-        state.addPoint(ramAccounting, 4L, 15.0);
-        state.addPoint(ramAccounting, 5L, 11.0);
-
-        List<Point> result = state.downsample();
-        assertThat(result).extracting(i -> i.timestamp).containsExactly(1L, 3L, 5L);
-        assertThat(result).extracting(i -> i.value).containsExactly(10.0, 50.0, 11.0);
-    }
-
-    @Test
-    public void test_should_select_the_points_that_form_largest_area_negative_datapoints_when_downsampling()
-            throws Exception {
-        LttbState state = new LttbState();
-        RamAccounting ramAccounting = new PlainRamAccounting();
-        state.init(ramAccounting, 3);
-
-        state.addPoint(ramAccounting, 1L, -10.0);
-        state.addPoint(ramAccounting, 2L, 12.0);
-        state.addPoint(ramAccounting, 3L, 50.0);
-        state.addPoint(ramAccounting, 4L, 15.0);
-        state.addPoint(ramAccounting, 5L, 11.0);
-
-        List<Point> result = state.downsample();
-        assertThat(result).extracting(i -> i.timestamp).containsExactly(1L, 3L, 5L);
-        assertThat(result).extracting(i -> i.value).containsExactly(-10.0, 50.0, 11.0);
-    }
-
-    @Test
-    public void test_should_not_downsample_rather_return_original_points_if_threshold_zero() throws Exception {
-        LttbState state = new LttbState();
-        RamAccounting ramAccounting = new PlainRamAccounting();
-        state.init(ramAccounting, 0);
-
-        state.addPoint(ramAccounting, 5L, 20.0);
-        state.addPoint(ramAccounting, 4L, 10.0);
-        state.addPoint(ramAccounting, 3L, 20.0);
-        state.addPoint(ramAccounting, 2L, 80.0);
-        state.addPoint(ramAccounting, 1L, 90.0);
-
-        assertThat(state.getPoints()).hasSize(5);
-
-        List<Point> result = state.downsample();
-        assertThat(result).extracting(i -> i.timestamp).containsExactly(1L, 2L, 3L, 4L, 5L);
-        assertThat(result).extracting(i -> i.value).containsExactly(90.0, 80.0, 20.0, 10.0, 20.0);
-    }
-
-    @Test
-    public void test_should_not_downsample_rather_return_original_points_if_threshold_larger_than_amount_of_data()
-            throws Exception {
-        LttbState state = new LttbState();
-        RamAccounting ramAccounting = new PlainRamAccounting();
-        state.init(ramAccounting, 10);
-
-        state.addPoint(ramAccounting, 5L, 20.0);
-        state.addPoint(ramAccounting, 4L, 10.0);
-        state.addPoint(ramAccounting, 3L, 20.0);
-        state.addPoint(ramAccounting, 2L, 80.0);
-        state.addPoint(ramAccounting, 1L, 90.0);
-
-        assertThat(state.getPoints()).hasSize(5);
-
-        List<Point> result = state.downsample();
-        assertThat(result).extracting(i -> i.timestamp).containsExactly(1L, 2L, 3L, 4L, 5L);
-        assertThat(result).extracting(i -> i.value).containsExactly(90.0, 80.0, 20.0, 10.0, 20.0);
     }
 
     @Test
@@ -175,64 +60,9 @@ public class LttbStateTest extends AggregationTestCase {
         assertThat(state1.isInitialized()).isEqualTo(state2.isInitialized());
 
         for (int i = 0; i < state1.getPoints().size(); i++) {
-            assertThat(state1.getPoints().get(i).timestamp).isEqualTo(state2.getPoints().get(i).timestamp);
-            assertThat(state1.getPoints().get(i).value).isEqualTo(state2.getPoints().get(i).value);
+            assertThat(state1.getPoints().get(i).timestamp()).isEqualTo(state2.getPoints().get(i).timestamp());
+            assertThat(state1.getPoints().get(i).value()).isEqualTo(state2.getPoints().get(i).value());
         }
-    }
-
-    @Test
-    public void test_should_merge_two_states() throws Exception {
-        LttbState state1 = new LttbState();
-        RamAccounting ramAccounting = new PlainRamAccounting();
-        state1.init(ramAccounting, 10);
-        state1.addPoint(ramAccounting, 20L, 22.0);
-
-        LttbState state2 = new LttbState();
-        RamAccounting ramAccounting2 = new PlainRamAccounting();
-        state2.init(ramAccounting2, 10);
-        state2.addPoint(ramAccounting2, 50L, 1.0);
-
-        state1.merge(ramAccounting, state2);
-
-        assertThat(state1.getPoints()).extracting(i -> i.timestamp).containsExactly(20L, 50L);
-        assertThat(state1.getPoints()).extracting(i -> i.value).containsExactly(22.0, 1.0);
-    }
-
-    @Test
-    public void test_raises_during_merge_if_states_are_inconsistent() throws Exception {
-        LttbState state1 = new LttbState();
-        RamAccounting ramAccounting = new PlainRamAccounting();
-        state1.init(ramAccounting, 10);
-        state1.addPoint(ramAccounting, 20L, 22.0);
-
-        LttbState state2 = new LttbState();
-        RamAccounting ramAccounting2 = new PlainRamAccounting();
-        state2.init(ramAccounting2, 200); // threshold does not match state1
-        state2.addPoint(ramAccounting2, 50L, 1.0);
-
-        assertThatThrownBy(() -> state1.merge(ramAccounting, state2))
-                .isExactlyInstanceOf(IllegalStateException.class)
-                .hasMessage("The thresholds between the two LttbStates do not match");
-    }
-
-    @Test
-    public void test_init_should_succeed_if_not_already_initialized() throws Exception {
-        LttbState state = new LttbState();
-        state.init(RamAccounting.NO_ACCOUNTING, 10);
-
-        assertThat(state.getPoints()).isNotNull();
-        assertThat(state.getThreshold()).isEqualTo(10);
-        assertThat(state.isInitialized()).isTrue();
-    }
-
-    @Test
-    public void test_init_raises_if_already_initialized() throws Exception {
-        LttbState state = new LttbState();
-        state.init(RamAccounting.NO_ACCOUNTING, 10);
-
-        assertThatThrownBy(() -> state.init(RamAccounting.NO_ACCOUNTING, 01))
-                .isExactlyInstanceOf(AssertionError.class)
-                .hasMessage("LttbState was already initialized");
     }
 
     @Test
@@ -253,7 +83,7 @@ public class LttbStateTest extends AggregationTestCase {
         state.addPoint(ramAccounting, 2L, 0.1);
         state.addPoint(ramAccounting, 3L, 0.1);
         assertThat(ramAccounting.totalBytes())
-                .isEqualTo(ramBefore + LargestTriangleThreeBucketsAggregation.Point.SHALLOW_SIZE * 3);
+                .isEqualTo(ramBefore + LTTBAggregation.Point.SHALLOW_SIZE * 3);
     }
 
 }
